JSON PHP Filter
DevBridge auto-sorts local json, but requires that remote json be
presorted. This file implements a json filter for remote json files. It
accepts multiple json files and converts them to a "suggestion" format.
A value key must still be provided in the json for DevBridge. JSON
records are sorted in ascending or descending order. The filter returns
all records that match the query from any value, not just the sorted
key.  More at:
http://itnachos.com/wp-admin/post.php?post=405&action=edit